requirements.txt
click==8.1.7
MarkupSafe==2.1.5
svgwrite==1.4.3
mysql-connector-python~=8.0.27
Jinja2~=3.0.2
python-dotenv~=0.19.1
loguru~=0.5.3